refactor(elasticsearch): Changed data_node_disk to data_node_disk_size and transfer code to service
