+++ alicloud/data_source_alicloud_elasticsearch_instances.go
@@ -1,6 +1,8 @@
 package alicloud
 
 import (
+	"regexp"
+
 	"github.com/aliyun/alibaba-cloud-sdk-go/sdk/requests"
 	"github.com/aliyun/alibaba-cloud-sdk-go/services/elasticsearch"
 	"github.com/hashicorp/terraform/helper/schema"
@@ -12,11 +14,12 @@ func dataSourceAlicloudElasticsearch() *schema.Resource {
 		Read: dataSourceAlicloudElasticsearchRead,
 
 		Schema: map[string]*schema.Schema{
-			"description": {
+			"description_regex": {
 				Type:     schema.TypeString,
 				Optional: true,
+				ValidateFunc: validateNameRegex,
 			},
-			"es_version": {
+			"version": {
 				Type:     schema.TypeString,
 				Optional: true,
 				ValidateFunc: validateAllowedStringValue([]string{
@@ -55,7 +58,7 @@ func dataSourceAlicloudElasticsearch() *schema.Resource {
 							Type:     schema.TypeString,
 							Computed: true,
 						},
-						"data_node_disk": {
+						"data_node_disk_size": {
 							Type:     schema.TypeInt,
 							Computed: true,
 						},
@@ -63,7 +66,7 @@ func dataSourceAlicloudElasticsearch() *schema.Resource {
 							Type:     schema.TypeString,
 							Computed: true,
 						},
-						"es_version": {
+						"version": {
 							Type:     schema.TypeString,
 							Computed: true,
 						},
@@ -91,9 +94,9 @@ func dataSourceAlicloudElasticsearchRead(d *schema.ResourceData, meta interface{
 
 	request := elasticsearch.CreateListInstanceRequest()
 	request.RegionId = client.RegionId
-	request.Description = d.Get("description").(string)
-	request.EsVersion = d.Get("es_version").(string)
+	request.EsVersion = d.Get("version").(string)
 	request.Size = requests.NewInteger(PageSizeLarge)
+	request.Page = requests.NewInteger(1)
 
 	var instances []elasticsearch.Instance
 
@@ -102,7 +105,7 @@ func dataSourceAlicloudElasticsearchRead(d *schema.ResourceData, meta interface{
 			return elasticsearchClient.ListInstance(request)
 		})
 		if err != nil {
-			return err
+			return BuildWrapError(request.GetActionName(), d.Id(), AlibabaCloudSdkGoERROR, err)
 		}
 		resp, _ := raw.(*elasticsearch.ListInstanceResponse)
 		if resp == nil || len(resp.Result) < 1 {
@@ -118,13 +121,32 @@ func dataSourceAlicloudElasticsearchRead(d *schema.ResourceData, meta interface{
 		}
 
 		if page, err := getNextpageNumber(request.Page); err != nil {
-			return err
+			return BuildWrapError(request.GetActionName(), d.Id(), AlibabaCloudSdkGoERROR, err)
 		} else {
 			request.Page = page
 		}
 	}
 
-	return extractInstance(d, instances)
+	var filteredInstance []elasticsearch.Instance
+	descriptionRegex, ok := d.GetOk("description_regex")
+	if (ok && descriptionRegex.(string) != "") || (len(instances) > 0) {
+		var r *regexp.Regexp
+		if descriptionRegex != "" {
+			r = regexp.MustCompile(descriptionRegex.(string))
+		}
+
+		for _, instance := range instances {
+			if r != nil && !r.MatchString(instance.Description) {
+				continue
+			}
+
+			filteredInstance = append(filteredInstance, instance)
+		}
+	} else {
+		filteredInstance = instances
+	}
+
+	return extractInstance(d, filteredInstance)
 }
 
 func extractInstance(d *schema.ResourceData, instances []elasticsearch.Instance) error {
@@ -138,10 +160,10 @@ func extractInstance(d *schema.ResourceData, instances []elasticsearch.Instance)
 			"instance_charge_type": getChargeType(item.PaymentType),
 			"data_node_amount":     item.NodeAmount,
 			"data_node_spec":       item.NodeSpec.Spec,
-			"data_node_disk":       item.NodeSpec.Disk,
+			"data_node_disk_size":  item.NodeSpec.Disk,
 			"data_node_disk_type":  item.NodeSpec.DiskType,
 			"status":               item.Status,
-			"es_version":           item.EsVersion,
+			"version":              item.EsVersion,
 			"created_at":           item.CreatedAt,
 			"updated_at":           item.UpdatedAt,
 			"vswitch_id":           item.NetworkConfig.VswitchId,
