--- examples/elasticsearch/output.tf
@@ -1,3 +1,0 @@
-output "elasticsearch_ids" {
-  value = "${join(",", alicloud_elasticsearch.instance.*.id)}"
-}
